echo el Guardar Detalle de Paquete

+++ database/migrations/2020_09_01_200553_create_paqueteitem_table.php
@@ -18,7 +18,7 @@ class CreatePaqueteitemTable extends Migration
             $table->foreignId('idPaquete');
             $table->foreignId('idItem');
             $table->integer('cantidad');
-            $table->float('precio',8,2);
+            $table->decimal('precio',20,2);
             
             $table->foreign('idPaquete')->references('id')->on('paquete')->ondelete('cascade');
             $table->foreign('idItem')->references('id')->on('item');
